reconstruct the package
